requirements.txt
hop-client >= 0.5.0
jsonschema
python-dotenv
click
tabulate